Make gaiacli and gaiad accessable through functions
