--- cmd/gaia/cmd/gaiacli/utils.go
@@ -1,197 +1,0 @@
-package main
-
-import (
-	"net/http"
-	"os"
-	"path"
-	"path/filepath"
-	"strings"
-
-	"github.com/cosmos/cosmos-sdk/client"
-	"github.com/cosmos/cosmos-sdk/client/keys"
-	"github.com/cosmos/cosmos-sdk/client/lcd"
-	"github.com/cosmos/cosmos-sdk/client/rpc"
-	"github.com/cosmos/cosmos-sdk/client/tx"
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	authcmd "github.com/cosmos/cosmos-sdk/x/auth/client/cli"
-	auth "github.com/cosmos/cosmos-sdk/x/auth/client/rest"
-	bankcmd "github.com/cosmos/cosmos-sdk/x/bank/client/cli"
-	bank "github.com/cosmos/cosmos-sdk/x/bank/client/rest"
-	gov "github.com/cosmos/cosmos-sdk/x/gov/client/rest"
-	slashing "github.com/cosmos/cosmos-sdk/x/slashing/client/rest"
-	stake "github.com/cosmos/cosmos-sdk/x/stake/client/rest"
-	"github.com/pkg/errors"
-	"github.com/rakyll/statik/fs"
-	"github.com/spf13/cobra"
-	"github.com/spf13/viper"
-	amino "github.com/tendermint/go-amino"
-)
-
-// registerRoutes registers the routes from the different modules for the LCD.
-// NOTE: details on the routes added for each module are in the module documentation
-// NOTE: If making updates here you also need to update the test helper in client/lcd/test_helper.go
-func registerRoutes(rs *lcd.RestServer) {
-	registerSwaggerUI(rs)
-	keys.RegisterRoutes(rs.Mux, rs.CliCtx.Indent)
-	rpc.RegisterRoutes(rs.CliCtx, rs.Mux)
-	tx.RegisterRoutes(rs.CliCtx, rs.Mux, rs.Cdc)
-	auth.RegisterRoutes(rs.CliCtx, rs.Mux, rs.Cdc, storeAcc)
-	bank.RegisterRoutes(rs.CliCtx, rs.Mux, rs.Cdc, rs.KeyBase)
-	stake.RegisterRoutes(rs.CliCtx, rs.Mux, rs.Cdc, rs.KeyBase)
-	slashing.RegisterRoutes(rs.CliCtx, rs.Mux, rs.Cdc, rs.KeyBase)
-	gov.RegisterRoutes(rs.CliCtx, rs.Mux, rs.Cdc)
-}
-
-func queryCmd(cdc *amino.Codec, mc []sdk.ModuleClients) *cobra.Command {
-	queryCmd := &cobra.Command{
-		Use:     "query",
-		Aliases: []string{"q"},
-		Short:   "Querying subcommands",
-	}
-
-	queryCmd.AddCommand(
-		rpc.ValidatorCommand(),
-		rpc.BlockCommand(),
-		tx.SearchTxCmd(cdc),
-		tx.QueryTxCmd(cdc),
-		client.LineBreak,
-		authcmd.GetAccountCmd(storeAcc, cdc),
-	)
-
-	for _, m := range mc {
-		queryCmd.AddCommand(m.GetQueryCmd())
-	}
-
-	return queryCmd
-}
-
-func txCmd(cdc *amino.Codec, mc []sdk.ModuleClients) *cobra.Command {
-	txCmd := &cobra.Command{
-		Use:   "tx",
-		Short: "Transactions subcommands",
-	}
-
-	txCmd.AddCommand(
-		bankcmd.SendTxCmd(cdc),
-		client.LineBreak,
-		authcmd.GetSignCommand(cdc),
-		bankcmd.GetBroadcastCommand(cdc),
-		client.LineBreak,
-	)
-
-	for _, m := range mc {
-		txCmd.AddCommand(m.GetTxCmd())
-	}
-
-	return txCmd
-}
-
-func registerSwaggerUI(rs *lcd.RestServer) {
-	statikFS, err := fs.New()
-	if err != nil {
-		panic(err)
-	}
-	staticServer := http.FileServer(statikFS)
-	rs.Mux.PathPrefix("/swagger-ui/").Handler(http.StripPrefix("/swagger-ui/", staticServer))
-}
-
-func prepareMainCmd(cmd *cobra.Command, envPrefix, defaultHome string) *cobra.Command {
-	initConfig(cmd)
-	cobra.OnInitialize(func() { initEnv(envPrefix) })
-	cmd.PersistentFlags().StringP(outputFlag, "o", "text", "Output format (text|json)")
-	cmd.PersistentFlags().String(homeFlag, defaultHome, "directory for config and data")
-	cmd.PersistentFlags().Bool(traceFlag, false, "print out full stack trace on errors")
-	cmd.PersistentPreRunE = concatCobraCmdFuncs(bindFlagsLoadViper, validateOutput, cmd.PersistentPreRunE)
-	return cmd
-}
-
-func initConfig(cmd *cobra.Command) error {
-	home, err := cmd.PersistentFlags().GetString(homeFlag)
-	if err != nil {
-		return err
-	}
-
-	cfgFile := path.Join(home, "config", "config.toml")
-	if _, err := os.Stat(cfgFile); err == nil {
-		viper.SetConfigFile(cfgFile)
-
-		if err := viper.ReadInConfig(); err != nil {
-			return err
-		}
-	}
-	return viper.BindPFlag(outputFlag, cmd.PersistentFlags().Lookup(outputFlag))
-}
-
-// initEnv sets prefix to use for ENV variables
-func initEnv(prefix string) {
-	prefix = strings.ToUpper(prefix)
-	ps := prefix + "_"
-	for _, e := range os.Environ() {
-		kv := strings.SplitN(e, "=", 2)
-		if len(kv) == 2 {
-			k, v := kv[0], kv[1]
-			if strings.HasPrefix(k, prefix) && !strings.HasPrefix(k, ps) {
-				k2 := strings.Replace(k, prefix, ps, 1)
-				os.Setenv(k2, v)
-			}
-		}
-	}
-
-	// env variables with prefix (eg. {{ .prefix }}_ROOT)
-	viper.SetEnvPrefix(prefix)
-	viper.SetEnvKeyReplacer(strings.NewReplacer(".", "_", "-", "_"))
-	viper.AutomaticEnv()
-}
-
-// cobraCmdFuncs allows for multiple PersistentPreRun hooks to run for each command
-type cobraCmdFunc func(cmd *cobra.Command, args []string) error
-
-// Returns a single function that calls each argument function in sequence
-// RunE, PreRunE, PersistentPreRunE, etc. all have this same signature
-func concatCobraCmdFuncs(fs ...cobraCmdFunc) cobraCmdFunc {
-	return func(cmd *cobra.Command, args []string) error {
-		for _, f := range fs {
-			if f != nil {
-				if err := f(cmd, args); err != nil {
-					return err
-				}
-			}
-		}
-		return nil
-	}
-}
-
-func validateOutput(cmd *cobra.Command, args []string) error {
-	// validate output format
-	output := viper.GetString(outputFlag)
-	switch output {
-	case "text", "json":
-	default:
-		return errors.Errorf("Unsupported output format: %s", output)
-	}
-	return nil
-}
-
-// Bind all flags and read the config into viper
-func bindFlagsLoadViper(cmd *cobra.Command, args []string) error {
-	// cmd.Flags() includes flags from this command and all persistent flags from the parent
-	if err := viper.BindPFlags(cmd.Flags()); err != nil {
-		return err
-	}
-
-	homeDir := viper.GetString(homeFlag)
-	viper.Set(homeFlag, homeDir)
-	viper.SetConfigName("config")                         // name of config file (without extension)
-	viper.AddConfigPath(homeDir)                          // search root directory
-	viper.AddConfigPath(filepath.Join(homeDir, "config")) // search root directory /config
-
-	// If a config file is found, read it in.
-	if err := viper.ReadInConfig(); err == nil {
-		// stderr, so if we redirect output to json file, this doesn't appear
-		// fmt.Fprintln(os.Stderr, "Using config file:", viper.ConfigFileUsed())
-	} else if _, ok := err.(viper.ConfigFileNotFoundError); !ok {
-		// ignore not found error, return other errors
-		return err
-	}
-	return nil
-}
